Move helper functions into utils
