+++ src/tracerbench/trace/index.js
@@ -8,4 +8,4 @@ export { loadTrace } from './load_trace';
 export { liveTrace } from './live_trace';
 export { networkConditions, IConditions, INetworkConditions, } from './conditions';
 export * from './trace_event';
-export { getBrowserArgs } from './utils';
+export { getBrowserArgs } from '../../util/utils';
